Moved anything that should not change year to year in to a package called core
