+++ src/main/java/ca/team2706/vision/core/Bundle.java
@@ -1,8 +1,8 @@
-package ca.team2706.vision.vision2019;
+package ca.team2706.vision.core;
 
 import java.awt.image.BufferedImage;
 
-import ca.team2706.vision.vision2019.params.VisionParams;
+import ca.team2706.vision.core.params.VisionParams;
 
 public class Bundle {
 	private BufferedImage raw,binMask,output;
