+++ src/main/java/ca/team2706/vision/core/ConfigParser.java
@@ -1,4 +1,4 @@
-package ca.team2706.vision.vision2019;
+package ca.team2706.vision.core;
 
 import java.io.File;
 import java.io.FileWriter;
