+++ src/main/java/ca/team2706/vision/core/VisionCameraServer.java
@@ -1,4 +1,4 @@
-package ca.team2706.vision.vision2019;
+package ca.team2706.vision.core;
 
 import java.util.HashMap;
 import java.util.Map;
