+++ src/test/java/ca/team2706/vision/vision2019/BundleTest.java
@@ -6,7 +6,7 @@ import java.awt.image.BufferedImage;
 
 import org.junit.Test;
 
-import ca.team2706.vision.vision2019.Bundle;
+import ca.team2706.vision.core.Bundle;
 
 public class BundleTest {
